+++ library/src/commonMain/kotlin/io/github/ronjunevaldoz/paymongo/serialization/PayMongoJson.kt
@@ -1,0 +1,11 @@
+package io.github.ronjunevaldoz.paymongo.serialization
+
+import kotlinx.serialization.json.Json
+
+val PayMongoJson = Json {
+    serializersModule = ResourceModule
+    prettyPrint = true
+    ignoreUnknownKeys = true
+    isLenient = true
+    prettyPrint = true
+}