+++ library/src/commonMain/kotlin/io/github/ronjunevaldoz/paymongo/util/PayMongoClientFactory.kt
@@ -1,10 +1,10 @@
 package io.github.ronjunevaldoz.paymongo.util
 
-import io.github.ronjunevaldoz.paymongo.Paymongo.Config
-import io.github.ronjunevaldoz.paymongo.PaymongoJson
+import io.github.ronjunevaldoz.paymongo.PayMongo.Config
 import io.github.ronjunevaldoz.paymongo.exception.PayMongoException
 import io.github.ronjunevaldoz.paymongo.models.error.PayMongoErrorResponse
 import io.ktor.client.HttpClient
+import io.ktor.client.call.body
 import io.ktor.client.plugins.ClientRequestException
 import io.ktor.client.plugins.HttpResponseValidator
 import io.ktor.client.plugins.RedirectResponseException
@@ -24,19 +24,14 @@ import io.ktor.client.plugins.observer.ResponseObserver
 import io.ktor.client.request.HttpRequest
 import io.ktor.client.request.header
 import io.ktor.client.statement.HttpResponse
-import io.ktor.client.statement.content
 import io.ktor.http.ContentType
 import io.ktor.http.HttpHeaders
 import io.ktor.http.contentType
 import io.ktor.serialization.kotlinx.json.json
-import io.ktor.utils.io.ByteReadChannel
-import io.ktor.utils.io.InternalAPI
-import io.ktor.utils.io.readUTF8Line
 
-object PaymongoClientFactory {
+object PayMongoClientFactory {
     private const val API_V1 = "https://api.paymongo.com/v1"
 
-    @OptIn(InternalAPI::class)
     fun client(
         config: Config
     ): HttpClient = HttpClient {
@@ -59,7 +54,9 @@ object PaymongoClientFactory {
             }
         }
         install(ResponseObserver) {
-            onResponse { response -> }
+            onResponse { response ->
+
+            }
         }
         HttpResponseValidator {
             validateResponse { response: HttpResponse ->
@@ -76,8 +73,8 @@ object PaymongoClientFactory {
             }
 
             handleResponseExceptionWithRequest { cause: Throwable, request: HttpRequest ->
-                val error = when (cause) {
-                    is ClientRequestException -> getError(cause.response.content)
+                val error: PayMongoErrorResponse = when (cause) {
+                    is ClientRequestException -> cause.response.body()
                     else -> throw cause
                 }
                 throw PayMongoException(error.errors)
@@ -88,13 +85,5 @@ object PaymongoClientFactory {
             contentType(ContentType.Application.Json)
             header(HttpHeaders.Accept, "application/json")
         }
-    }
-
-
-    private suspend fun getError(responseContent: ByteReadChannel): PayMongoErrorResponse {
-        responseContent.readUTF8Line()?.let {
-            return PaymongoJson.decodeFromString(it)
-        }
-        throw IllegalArgumentException("not a parsable error")
     }
 }